+++ server/pubsub/src/pubsub/collection.h
@@ -21,7 +21,7 @@ typedef struct {
 
 /* Protocol-specific identifiers for PubSub */
 typedef struct {
-    uint32_t publisherId;
+    uint16_t publisherId;
     uint16_t groupId;
     uint16_t writerId;
     uint16_t nsIndex;
